+++ interpreter/k8s/src/test/scala/io/buoyant/transformer/k8s/LocalnodeTransformerTest.scala
@@ -4,9 +4,8 @@ import com.twitter.finagle.Name.Bound
 import com.twitter.finagle._
 import com.twitter.finagle.naming.NameInterpreter
 import com.twitter.util.{Activity, Future, Var}
-import io.buoyant.namer.{Metadata, MetadataFiltertingNameTreeTransformer}
+import io.buoyant.namer.{Metadata, MetadataFiltertingNameTreeTransformer, Netmask, SubnetLocalTransformer}
 import io.buoyant.test.FunSuite
-import io.buoyant.transformer.{Netmask, SubnetLocalTransformer}
 import java.net.{InetAddress, InetSocketAddress}
 
 class LocalnodeTransformerTest extends FunSuite {